[SMLRCreateAccountViewController] Code Cleanup: fixed compiler warning during archieve

+++ Simlar/src/SMLRCreateAccountViewController.m
@@ -104,7 +104,7 @@
                         return;
                     default:
                         [SMLRCreateAccountViewController showErrorAlertWithTitle:@"Server Error"
-                                                                         message:[NSString stringWithFormat:@"Internal Error with number: %i", error.code]];
+                                                                         message:[NSString stringWithFormat:@"Internal Error with number: %li", (long)error.code]];
                         return;
                 }
             }
